Runs, but run out of space?

+++ main/audio/I2SOutput.cpp
@@ -10,12 +10,12 @@ I2SOutput::I2SOutput(i2s_port_t i2s_port, i2s_pin_config_t &i2s_pins) : Output(i
 }
 
 
-void I2SOutput::start(int sample_rate)
+void I2SOutput::start(uint32_t sample_rate)
 {
     // i2s config for writing both channels of I2S
     i2s_config_t i2s_config = {
         .mode = (i2s_mode_t)(I2S_MODE_MASTER | I2S_MODE_TX),
-        .sample_rate = (uint32_t)sample_rate,
+        .sample_rate = sample_rate,
         .bits_per_sample = I2S_BITS_PER_SAMPLE_16BIT,
         .channel_format = I2S_CHANNEL_FMT_RIGHT_LEFT,
         .communication_format = I2S_COMM_FORMAT_STAND_I2S,
